cdk/test/test-requirements.txt
strands-agents>=0.1.6
strands-agents-tools>=0.1.4
pytest
